2020.02.26 가입한상품(완료) 장바구니(진행중)

+++ src/main/webapp/WEB-INF/views/board_Mypage/myInfo.jsp
@@ -11,8 +11,9 @@ contentType="text/html; charset=EUC-KR" pageEncoding="EUC-KR"%>
 <link href="http://fonts.googleapis.com/css?family=Source+Sans+Pro:200,300,400,600,700,900" rel="stylesheet" />
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/default.css" >
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/fonts.css" >
+<link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/product.css" >
 <meta http-equiv="Content-Type" content="text/html; charset=EUC-KR">
-<title>Insert title here</title>
+<title>아워뱅크::내정보</title>
 <script>
 function goToMyInfoUpdate(){
 	location.href='myInfoUpdateForm.do?id=${userBean.getId()}';
@@ -21,8 +22,14 @@ function goToMyPge(){
 	location.href='myPage.do';
 }
 function deleteId(){
-	location.href='deleteId.do?id=${userBean.getId()}';
+	if(confirm('정말 삭제하시겠습니까?')==true){
+		location.href='deleteId.do?id=${userBean.getId()}';
+	}else {
+		return;
+	}
+	
 }
+
 </script>
 </head>
 
@@ -46,7 +53,7 @@ function deleteId(){
 	
 <!-- *********************** 내용 ****************************  -->
 <div id="my_wrap">
-	<table border="1" cellpadding="10px" >
+	<table border="1" cellpadding="10px" c>
 		<tr>
 		<td>
 			<ul>
